refactor: put characters classes into module
