+++ character/playable.rb
@@ -1,0 +1,4 @@
+class Playable < Character
+
+end
+
