+++ character/stats.rb
@@ -1,0 +1,13 @@
+class Stats < Value
+    def initialize(hit_point, magic_point, attack)
+        @max_hit_points = max_hit_points
+        @max_magic_points = max_magic_points
+    end
+
+
+    def max_hit_points=(value)
+
+    end
+end
+
+
